refactor: extract log search service and update logs contracts
- Add `headers: authHeadersSchema` to `logsListContract` for CLI token auth
- Define `zeroLogsSearchContract` in zero-runs.ts for `/api/zero/logs/search`
- Export new contract from `@vm0/core`
- Extract search helpers into shared `log-search-service.ts`
- Refactor existing search route to import from shared service

Closes #7634
